[MemAlloc] Fix 64b

+++ src/Kernel/Math/TRUSTTab_parts.tpp
@@ -78,7 +78,7 @@ inline void init_parts(TRUSTVect<_TYPE_>& vect, TRUST_Vector<TRUSTTab<_TYPE_>>& 
           shape_i[0] = md_part.valeur().get_nb_items_tot();
           TRUSTTab<_TYPE_>& part = parts[i];
           // Fait pointer le domaine de memoire sur le sous-tableau (pour l'instant tableau monodimensionnel)
-          part.ref_array(vect, offset * line_size, shape_i[0] * line_size * std::max(mdata.shapes_[i], 1));
+          part.ref_array(vect, offset * line_size, shape_i[0] * line_size * std::max(mdata.shapes_[i], (int)1)); // int cast necessary for 64b transcription
           // Change le "shape" du tableau pour mettre le nombre de lignes et de colonnes
           // (nombre total d'items inchange, donc resize autorise)
           part.resize(shape_i);
